reduce the go type of number columns to just two: 'int' for incremental IDs, sizes, number of decimals or progresses, and 'int64' for block numbers, balances, chainIDs

+++ api/strategy.go
@@ -45,7 +45,7 @@ func (capi *census3API) createDummyStrategy(tokenID []byte) error {
 		return err
 	}
 	_, err = capi.db.QueriesRW.CreateStrategyToken(ctx, queries.CreateStrategyTokenParams{
-		StrategyID: strategyID,
+		StrategyID: int(strategyID),
 		TokenID:    tokenID,
 		MinBalance: big.NewInt(0).Bytes(),
 		MethodHash: []byte("test"),
@@ -72,9 +72,9 @@ func (capi *census3API) getStrategies(msg *api.APIdata, ctx *httprouter.HTTPCont
 		return ErrNoStrategies
 	}
 	// parse and encode the strategies
-	strategies := GetStrategiesResponse{Strategies: []uint64{}}
+	strategies := GetStrategiesResponse{Strategies: []int{}}
 	for _, strategy := range rows {
-		strategies.Strategies = append(strategies.Strategies, uint64(strategy.ID))
+		strategies.Strategies = append(strategies.Strategies, strategy.ID)
 	}
 	res, err := json.Marshal(strategies)
 	if err != nil {
@@ -96,7 +96,7 @@ func (capi *census3API) getStrategy(msg *api.APIdata, ctx *httprouter.HTTPContex
 	// get strategy from the database
 	internalCtx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 	defer cancel()
-	strategyData, err := capi.db.QueriesRO.StrategyByID(internalCtx, int64(strategyID))
+	strategyData, err := capi.db.QueriesRO.StrategyByID(internalCtx, strategyID)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return ErrNotFoundStrategy.WithErr(err)
@@ -105,7 +105,7 @@ func (capi *census3API) getStrategy(msg *api.APIdata, ctx *httprouter.HTTPContex
 	}
 	// parse strategy information
 	strategy := GetStrategyResponse{
-		ID:        uint64(strategyData.ID),
+		ID:        strategyData.ID,
 		Predicate: strategyData.Predicate,
 		Tokens:    []GetStrategyToken{},
 	}
@@ -151,9 +151,9 @@ func (capi *census3API) getTokenStrategies(msg *api.APIdata, ctx *httprouter.HTT
 		return ErrNoStrategies
 	}
 	// parse and encode strategies
-	strategies := GetStrategiesResponse{Strategies: []uint64{}}
+	strategies := GetStrategiesResponse{Strategies: []int{}}
 	for _, tokenStrategy := range rows {
-		strategies.Strategies = append(strategies.Strategies, uint64(tokenStrategy.ID))
+		strategies.Strategies = append(strategies.Strategies, tokenStrategy.ID)
 	}
 	res, err := json.Marshal(strategies)
 	if err != nil {
